add terrain test; fix bugs

+++ examples/everlon/packages/world/src/systems/ActivateSystem.sol
@@ -28,10 +28,15 @@ contract ActivateSystem is ActivateProtoSystem {
   ) internal override {
     super.postEvent(actingObjectEntityId, objectTypeId, coord, eventEntityId, eventData);
 
-    // Clear all keys in Metadata
-    bytes32[][] memory objectsRan = getKeysInTable(MetadataTableId);
-    for (uint256 i = 0; i < objectsRan.length; i++) {
-      Metadata.deleteRecord(objectsRan[i][0]);
+    address callerAddress = _msgSender();
+    // Clear all keys in Metadata if not called by World or Simulator
+    // This would typically represent the end of a user call, vs the end of
+    // an internal call
+    if (callerAddress != _world() && callerAddress != getSimulatorAddress()) {
+      bytes32[][] memory objectsRan = getKeysInTable(MetadataTableId);
+      for (uint256 i = 0; i < objectsRan.length; i++) {
+        Metadata.deleteRecord(objectsRan[i][0]);
+      }
     }
   }
